+++ zcash_client_sqlite/src/wallet.rs
@@ -2365,7 +2365,7 @@ pub(crate) fn get_funding_accounts(
         conn,
         tx.transparent_bundle()
             .iter()
-            .flat_map(|bundle| bundle.vin.iter().map(|txin| &txin.prevout)),
+            .flat_map(|bundle| bundle.vin.iter().map(|txin| txin.prevout())),
     )?);
 
     funding_accounts.extend(sapling::detect_spending_accounts(
@@ -3434,7 +3434,7 @@ fn determine_fee(
                 for t_in in &t_bundle.vin {
                     match (
                         result,
-                        get_wallet_transparent_output(_conn, &t_in.prevout, true)?,
+                        get_wallet_transparent_output(_conn, t_in.prevout(), true)?,
                     ) {
                         (Some(b), Some(out)) => {
                             result = Some((b + out.txout().value).ok_or(BalanceError::Overflow)?)
@@ -3740,7 +3740,7 @@ pub(crate) fn store_decrypted_tx<P: consensus::Parameters>(
         .iter()
         .flat_map(|b| b.vin.iter())
     {
-        transparent::mark_transparent_utxo_spent(conn, tx_ref, &txin.prevout)?;
+        transparent::mark_transparent_utxo_spent(conn, tx_ref, txin.prevout())?;
     }
 
     #[cfg(feature = "transparent-inputs")]
@@ -4018,7 +4018,7 @@ pub(crate) fn queue_transparent_input_retrieval<AccountId>(
             // queue the transparent inputs for enhancement
             queue_tx_retrieval(
                 conn,
-                b.vin.iter().map(|txin| *txin.prevout.txid()),
+                b.vin.iter().map(|txin| *txin.prevout().txid()),
                 Some(tx_ref),
             )?;
         }