+++ zcash_primitives/src/transaction/tests.rs
@@ -262,10 +262,12 @@ fn zip_0244() {
                 vin: b
                     .vin
                     .iter()
-                    .map(|vin| TxIn {
-                        prevout: vin.prevout.clone(),
-                        script_sig: vin.script_sig.clone(),
-                        sequence: vin.sequence,
+                    .map(|vin| {
+                        TxIn::from_parts(
+                            vin.prevout().clone(),
+                            vin.script_sig().clone(),
+                            vin.sequence(),
+                        )
                     })
                     .collect(),
                 vout: b.vout.clone(),
@@ -421,10 +423,12 @@ fn zip_0233() {
                 vin: b
                     .vin
                     .iter()
-                    .map(|vin| TxIn {
-                        prevout: vin.prevout.clone(),
-                        script_sig: vin.script_sig.clone(),
-                        sequence: vin.sequence,
+                    .map(|vin| {
+                        TxIn::from_parts(
+                            vin.prevout().clone(),
+                            vin.script_sig().clone(),
+                            vin.sequence(),
+                        )
                     })
                     .collect(),
                 vout: b.vout.clone(),