+++ zcash_transparent/CHANGELOG.md
@@ -10,6 +10,14 @@ workspace.
 
 ## [Unreleased]
 
+### Added
+- `zcash_transparent::bundle::TxIn::from_parts`
+- `zcash_transparent::bundle::TxIn::{prevout, script_sig, sequence}` accessor methods.
+
+### Deprecated
+- `zcash_transparent::bundle::TxIn::{prevout, script_sig, sequence}` public fields
+  are deprecated; use the newly added accessor methods instead.
+
 ## [0.4.0] - 2025-07-31
 
 ### Added